Add files via upload
This thesis focused on the development of in-house computer program for the analysis of infrared (IR) thermal images. Infrared thermography was chosen as it is a non-destructive inspection technique for detecting defects in carbon fiber reinforced resin composites. In this thesis, firstly, an introduction to the fundamental concepts of infrared thermography and digital image processing and analysis is made. Various algorithms for processing thermographic data are then developed and tested for their effectiveness and detection capabilities. Four thermographic image analysis algorithms were applied to material samples with three different heating times. An infrared camera recorded the temperature field and produced a sequence of thermal images which in a post processing stage were analyzed to find faults-defects. These defects were successfully detected. Lock-in thermography proved to be an accurate method and gave better results compared to Pulse thermography. It should be stressed out that the numerical computations associated with the IR Lock-in thermography were the most time-consuming in terms of computational time.